+++ python/src/gherkin/token.py
@@ -10,8 +10,5 @@ class Token:
     def eof(self) -> bool:
         return not self.line
 
-    def detach(self) -> None:
-        pass  # TODO: detach line - is this needed?
-
     def token_value(self) -> str:
         return "EOF" if self.eof() else self.line.get_line_text()